requirements.txt
nextcord==2.3.2
python-dotenv==1.0.0
certifi==2023.7.22
requests==2.31.0
gradio-client==0.5.1